Assistant URL in assistant.ini i obsolete
The current WA_ENDPOINT is obsolete and needs to be updated. https://gateway.watsonplatform.net is obsolete.

Ideally it would be a documented parameter so that this can work with Assistant in any region.  I just hardcoded the us-south endpoint, but this change is untested. The URL should be a parameter so that it can be copied from the service and will work in any region.

I would have opened an issue instead of a PR, but this repo is not accepting issues.  So, I'm thinking this is unmaintained and should be archived.